altered entities, added security

+++ src/main/java/grsu/by/entity/Role.java
@@ -27,7 +27,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Short id;
 
     @Column(name = "name")
     private String name;
